refactor: fix reexports from index.ts files
