+++ src/hooks/index.ts
@@ -1,8 +1,9 @@
-export { default as useMediaQuery } from './useMediaQuery';
-export { default as useThrottle } from './useThrottle';
-export { default as useEvent } from './useEvent';
-export { default as useOutsideDetect } from './useOutsideDetect';
-export { default as useCursorLeave } from './useCursorLeave';
 export { default as useBackdrop } from './useBackdrop';
+export { default as useCursorLeave } from './useCursorLeave';
 export { default as useDebounce } from './useDebounce';
 export { default as useDisableScroll } from './useDisableScroll';
+export { default as useEvent } from './useEvent';
+export { default as useMediaQuery } from './useMediaQuery';
+export { default as useOutsideDetect } from './useOutsideDetect';
+export { default as useScroll } from './useScroll';
+export { default as useThrottle } from './useThrottle';
