+++ src/utils/index.ts
@@ -1,7 +1,10 @@
-export { default as modifyColor } from './colorModifier';
-export type { StringMap } from './TypeUtils';
+export { axios, axiosAuth } from './axios';
+export { default as bytesToFile } from './bytesToFile';
 export { default as calculateBackgroundColor } from './calculateBackgroundColor';
-export { default as createAnimation } from './createAnimation';
-export { parseLengthAndUnit, cssValue } from './unitConverter';
-export { default as debounce } from './debounce';
+export { default as capitalizedWords } from './capitalizedWords';
 export { default as checkIfServiceKey } from './checkIfServiceKey';
+export { default as debounce } from './debounce';
+export { parseLengthAndUnit, cssValue } from './unitConverter';
+export { default as modifyColor } from './modifyColor';
+export type { StringMap } from './TypeUtils';
+export { default as createAnimation } from './createAnimation';
